Re-merge autocompletion tests
